requirements.txt
gspread==5.10.0
oauth2client==4.1.3
pandas==2.0.3
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
rapidfuzz==3.2.0
